Share fixtures for creating and removing records
