+++ tests/conf_util.py
@@ -112,6 +112,62 @@ def session():
     return session
 
 
+@pytest.fixture()
+def load_bib(session, request):
+    return load(session, request)
+
+
+@pytest.fixture(scope='module')
+def load_bib_for_module(session, request):
+    return load(session, request)
+
+
+def load(session, request):
+    bib_ids = []
+
+    def load_bib(bib_file=BIB_FILE):
+        bib_id = create_bib(session=session, bib_file=bib_file)
+        bib_ids.append(bib_id)
+        trigger_elastic_refresh(session)
+        return bib_id
+
+    # Cleanup
+    def fin():
+        for bib_id in bib_ids:
+            result = delete_record(session, bib_id)
+            assert result.status_code == 204
+
+            result = session.get(bib_id)
+            assert result.status_code == 410
+
+    request.addfinalizer(fin)
+    return load_bib
+
+
+@pytest.fixture()
+def load_holding(session, request):
+    holding_ids = []
+
+    def load_holding(item_of=None):
+        holding_id = create_holding(session=session, item_of=item_of)
+        holding_ids.append(holding_id)
+        trigger_elastic_refresh(session)
+        return holding_id
+
+    # Cleanup
+    def fin():
+        for holding_id in holding_ids:
+            result = delete_record(session, holding_id)
+            assert result.status_code == 204
+
+            result = session.get(holding_id)
+            assert result.status_code == 410
+
+    request.addfinalizer(fin)
+
+    return load_holding
+
+
 def _get_input_value(page, name):
     xpath = page.xpath("//input[@name='{0}']/@value".format(name))
     return list(set(xpath))[0]
@@ -151,7 +207,7 @@ def put_post(session, thing_id, **kwargs):
     return session.put(thing_id, headers=headers, **kwargs)
 
 
-def delete_post(session, thing_id, **kwargs):
+def delete_record(session, thing_id, **kwargs):
     headers = {XL_ACTIVE_SIGEL_HEADER: ACTIVE_SIGEL}
     # Ensure records are present in the index before trying to delete them
     session.post(ES_REFRESH_URL)
@@ -204,3 +260,9 @@ def update_holding(session, holding_id, payload, etag):
                          data=json_payload,
                          headers=headers)
     return result
+
+
+def trigger_elastic_refresh(session):
+    result = session.post(ES_REFRESH_URL)
+    assert result.status_code == 200
+
